+++ main
@@ -1,68 +1,44 @@
 #!/usr/bin/env ruby
-require_relative 'actions'
-
-class App
-  include Actions
-
-  def initialize
-    @books = []
-    @rentals = []
-    @people = []
-  end
-
-  def start
-    puts 'Welcome to School Library App!'
-
-    loop do
-      actions
-
-      option = gets.chomp
-
-      break if option == '7'
-
-      handle_action option
-    end
-
-    puts 'Thank you for using this app!'
-  end
-
-  private
-
-  def handle_action(option)
-    case option
-    when '1'
-      list_books
-    when '2'
-      list_people
-    when '3'
-      create_person
-    when '4'
-      create_book
-    when '5'
-      create_rental
-    when '6'
-      list_rentals
-    else
-      puts 'That is not a valid option'
-    end
-  end
-
-  def actions
-    puts
-    puts 'Please choose an option by entering a number:'
-    puts '1 - List all books'
-    puts '2 - List all people'
-    puts '3 - Create a person'
-    puts '4 - Create a book'
-    puts '5 - Create a rental'
-    puts '6 - List all rentals for a given person id'
-    puts '7 - Exit'
-  end
-end
+require "./app.rb"
 
 def main
-  app = App.new
-  app.start
+  app = App.new()
+
+  response = nil
+
+  puts "Welcome to School Library App!", " ", " "
+  while response != "7"
+    puts "Please choose an option by enterin a number:"
+    puts "1 - List all books"
+    puts "2 - List all people"
+    puts "3 - Create a person"
+    puts "4 - Create a book"
+    puts "5 - Create a rental"
+    puts "6 - List all rentals for a given person id"
+    puts "7 - Exit"
+    response = gets.chomp
+
+    case response
+    when "1"
+      app.list_books
+    when "2"
+      app.list_people
+    when "3"
+      app.create_person
+    when "4"
+      app.create_book
+    when "5"
+      app.create_rental
+    when "6"
+      app.list_rentals
+    when "7"
+      puts "Thank you for using this app!"
+    end
+
+    puts "\n"
+  end
 end
+
+main()
 
 main
